+++ packages/keyring-snap-bridge/src/options.test.ts
@@ -1,0 +1,61 @@
+import type { SnapKeyringInternalOptions } from './options';
+import { getDefaultInternalOptions, getInternalOptionsOf } from './options';
+
+describe('options', () => {
+  const defaults = getDefaultInternalOptions();
+
+  describe('getInternalOptionsOf', () => {
+    it('returns default internal options if none are passed', () => {
+      expect(getInternalOptionsOf([])).toStrictEqual(defaults);
+    });
+
+    it('uses the first defined values', () => {
+      const input: Required<SnapKeyringInternalOptions> = {
+        displayConfirmation: !defaults.displayConfirmation,
+        displayAccountNameSuggestion: !defaults.displayAccountNameSuggestion,
+        setSelectedAccount: !defaults.setSelectedAccount,
+      };
+
+      expect(getInternalOptionsOf([input])).toStrictEqual(input);
+    });
+
+    it('uses all defined values from every items', () => {
+      const inputs: SnapKeyringInternalOptions[] = [
+        {
+          displayConfirmation: false,
+        },
+        {
+          displayAccountNameSuggestion: false,
+        },
+        {
+          setSelectedAccount: false,
+        },
+      ];
+
+      expect(getInternalOptionsOf(inputs)).toStrictEqual({
+        ...inputs[0],
+        ...inputs[1],
+        ...inputs[2],
+      });
+    });
+
+    it('fallbacks to the default options if one field is still not defined', () => {
+      const inputs: SnapKeyringInternalOptions[] = [
+        {
+          displayConfirmation: false,
+        },
+        // `displayAccountNameSuggestion` will be undefined
+        {
+          setSelectedAccount: false,
+        },
+      ];
+
+      expect(getInternalOptionsOf(inputs)).toStrictEqual({
+        ...inputs[0],
+        ...inputs[1],
+        // This one is defaulting to its default value
+        displayAccountNameSuggestion: defaults.displayAccountNameSuggestion,
+      });
+    });
+  });
+});